Move templated to their own folder
