+++ app/templates/DonutChart/DonutChartComponent.jsx
